Templates of template encode correctly however ...
Template is an interface that is public, which means someone could
technically create their own. Doing so would be kind of challenging, but
it could be done. Therfore the encoding and decoding stuff, because it
assumes the type of the encoded interface, likely should not live within
the template interface, and should instead be an additional feature of
the fpdf template.